* dispnew.c (Fframe_of_buffer_changed_p): Use ptrdiff_t, not int.

+++ src/dispnew.c
@@ -6033,7 +6033,7 @@ pass nil for VARIABLE.  */)
 {
   Lisp_Object state, tail, frame, buf;
   Lisp_Object *vecp, *end;
-  int n;
+  ptrdiff_t n;
 
   if (! NILP (variable))
     {
